% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chrVec}
\alias{chrVec}
\title{chromosome Vector}
\format{
A vector containing the chromosomes of the loci in RFQTLgenotype
}
\source{
unpublished data.
}
\usage{
chrVec
}
\description{
chromosome Vector
}
\keyword{datasets}
