% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{combineForestsExtended}
\alias{combineForestsExtended}
\title{Combine RFs}
\usage{
combineForestsExtended(...)
}
\arguments{
\item{...}{Two or more objects of class randomForest, to be combined into one.}
}
\value{
An object of class randomForest.
}
\description{
Function that combines several randomForest objects 
to one big Random Forest. Workaround for a bug in the
function  \code{combine} in \code{RandomForestExtended}.
}
