% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{cScore}
\alias{cScore}
\title{Combined RF score}
\usage{
cScore(rf = NULL, predictorsOfInterest = NULL, normalize = TRUE)
}
\arguments{
\item{rf}{A randomForest object that contains both permutation 
importance (PI) and the increase in node purity (RSS). Random Forests
that were generated with importance=TRUE are suited for this.}

\item{predictorsOfInterest}{Predictors for which the combined 
score should be computed. As a default all predictors are selected.}

\item{normalize}{Should the scores be normalized with the trait variance?
This makes the importance scores comparable across different traits.}
}
\value{
The output consists of a numeric vector containing a score 
 for each predictor of interest. The score is the product of RSS and PI.
 Negative RSS- or PI-scores are set to zero first. These scores are
 influenced by correlation between predictors and therefore contain
 a marker-specific bias.
}
\description{
Extracts a combined importance score from a randomForest object.
}
