% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{emma.kinship}
\alias{emma.kinship}
\title{emma kinship}
\usage{
emma.kinship(snps, method = "additive", use = "all")
}
\description{
Gratefully taken from emma 1.1.2 (url:http://mouse.cs.ucla.edu/emma/news.html,license: LGPL)
Used to generate thepopulation structure covariates for RF QTL mapping.
}
