% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{extractNAs}
\alias{extractNAs}
\title{Allele frequencies and missing genotypes.}
\usage{
extractNAs(genotype)
}
\arguments{
\item{genotype}{Genotype matrix with one row per individual, one
column per marker. Entries should be binary marker information (1,0,NA).}
}
\value{
List with one vector for each marker. The first element 
 is the frequency of allele 1. The following numbers indicate the
 indices of any missing values for the respective samples.
}
\description{
Function that calculates the allele frequencies (of the possible 
alleles 1 or 0) for each marker (column) and identifies missing values.
}
