% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{joinConsecutive}
\alias{joinConsecutive}
\title{Group Markers}
\usage{
joinConsecutive(QTLmat, chrVec)
}
\arguments{
\item{QTLmat}{Matrix consisting of two columns (target, predictor)
and one row per linked predictor.}

\item{chrVec}{Character-vector containing the chromosome on which 
a given marker is located.}
}
\value{
List with one entry for each regulated trait. Each entry 
 contains the target and a matrix with the start and end of the QTL-regions.
}
\description{
Function that joins consecutive markers that were significantly
linked to the same trait to a QTL.
}
