% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{joinCorrelated}
\alias{joinCorrelated}
\title{Join correlated QTL}
\usage{
joinCorrelated(QTLlist, corMat, corThreshold)
}
\arguments{
\item{QTLlist}{Output of joinNear.}

\item{corMat}{Squared matrix of predictor-correlation coefficients.}

\item{corThreshold}{Threshold indicating how strong QTL-regions have to
be correlated to be joined.}
}
\value{
List with an entry for each QTL. Each entry contains the 
 target and a matrix with the start and end of the QTL-regions after
 grouping highly correlated distant QTL.
}
\description{
Loin QTL-regions that are distant from each other but contain
highly correlated markers to QTL-groups using hierarchical clustering.
}
