% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{joinNear}
\alias{joinNear}
\title{Join close QTL}
\usage{
joinNear(QTLlist, corMat, corThreshold, distThreshold, chrVec)
}
\arguments{
\item{QTLlist}{Output from joinConsecutive.}

\item{corMat}{Squared matrix of predictor-correlation coefficients.}

\item{corThreshold}{Threshold indicating how strong QTL-regions have
to be correlated to be joined.}

\item{distThreshold}{Threshold indicating how close QTL-regions 
have to be to be joined.}

\item{chrVec}{Character vector containing the information on which
chromosome a marker is located.}
}
\value{
List with an entry for each regulated trait. Each entry 
 contains the target and a matrix with the start and end of the 
 QTL regions after combining close and correlated QTL.
}
\description{
Function that joins different QTL to a single one if they 
contain highly correlated markers.
}
