% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markerPositions}
\alias{markerPositions}
\title{chromosomal positions of the markers in RFQTLgenotype}
\format{
A dataframe with three columns, each
row corresponding to one locus.
\describe{
  \item{chr}{chromosome number}
  \item{start}{starting position of locus in bp}
  \item{end}{end position of locus in bp}
}
}
\source{
unpublished data.
}
\usage{
markerPositions
}
\description{
chromosomal positions of the markers in RFQTLgenotype
}
\keyword{datasets}
