% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{minPV}
\alias{minPV}
\title{Lead Marker}
\usage{
minPV(QTLlist, pmat)
}
\arguments{
\item{QTLlist}{List-object with an entry for each QTL as returned
by joinCorrelated.}

\item{pmat}{Matrix containing the p-values for each predictor (columns) 
and trait (rows).}
}
\value{
Object similar to QTLlist with added information regarding the most 
 significant predictor in each QTL.
}
\description{
Identify the predictor with the smallest p-value in each QTL.
}
