% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{perm1}
\alias{perm1}
\title{pre-computed permuted RF scores}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 688 rows and 10 columns.
}
\source{
unpublished data.
}
\usage{
perm1
}
\description{
A matrix containing null distribution of importance scores
generated through permutations.
}
\keyword{datasets}
