% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{permutedScores}
\alias{permutedScores}
\title{pre-computed permuted RF scores}
\format{}
\source{
unpublished data.
}
\usage{
permutedScores
}
\description{
A matrix containing null distribution of importance scores
generated through permutations.
}
\keyword{datasets}
