% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{pEst}
\alias{pEst}
\title{Estimate empirical p-value}
\usage{
pEst(path, markersPerIteration, scores, printProg = T, pCorrection = "none")
}
\arguments{
\item{path}{File path where the .RData files with the matrices of 
empirical values can be found. Thse have to conform to predictors 
(rows) X permutations (columns).}

\item{markersPerIteration}{Total number of null distributions that 
are loaded at the same time. The possible number is dependent on 
available working memory and size of the null distributions.}

\item{scores}{Real predictor-scores to be compared to the empirical 
null distributions. If scores is a matrix, it has to conform to 
traits (rows) X predictors (columns). Scores for the same predictor
in matrix, it has to conform to traits (rows) X predictors (columns).
Scores for the same predictor in different traits (rows) are 
compared to the same null distribution.}

\item{printProg}{Logical; if TRUE the last finished predictor is printed.}

\item{pCorrection}{Which type, if any, of multiple testing correction.
should be performed? Has to be either "none", "fdr" or "bonferroni".}
}
\value{
A matrix with empirical p-values is returned.
}
\description{
Function that uses finished permutations and real scores 
to estimate an empirical p-value.
}
