% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{readQTL}
\alias{readQTL}
\title{Load .qtl file}
\usage{
readQTL(path)
}
\arguments{
\item{path}{The location of the .qtl file that should be loaded.}
}
\value{
A list-object as it is returned by QTLgrouper.
}
\description{
Reconstructs a QTLlist from a .qtl file.
}
