% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{replaceGenoNAs}
\alias{replaceGenoNAs}
\title{Impute missing genotypes}
\usage{
replaceGenoNAs(genotype, NAlist)
}
\arguments{
\item{genotype}{Genotype as a matrix with 1,0,NA.}

\item{NAlist}{List object with one vector per predictor, containing
the frequency of the 1-allele, followed by the indices of 
individuals with missing values. As created by preMap.}
}
\value{
Matrix with the same dimensions as the input-genotype.
 NAs are replaced with alleles (0,1) according to the allele
 frequencies in the input (NAlist).
}
\description{
Function that replaces missing genotypes according to the
local allele frequencies.
}
