% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{RFepistasis}
\alias{RFepistasis}
\title{RF epistasis detection}
\usage{
RFepistasis(
  mappingData,
  markerInds1 = NULL,
  markerInds2 = NULL,
  pairMat = NULL,
  mtry = NULL,
  ntree = 30000,
  npermut = 100,
  nodesize = 5,
  minTest = 5,
  nthreads = 1,
  fullTable = F,
  Rcutoff = 0.9
)
}
\arguments{
\item{mappingData}{Dataset that was prepared with the function preMap.}

\item{markerInds1, markerInds2}{The indices of the markers that should 
be tested for interaction. All markers in markerInds1 will be tested 
against all markers in markerInds2. Will be ignored if pairMat is submitted.}

\item{pairMat}{An integer matrix with two columns indicating the
marker pairs that should be tested for interaction. Each row should
be the indices of the two markers that should be tested.}

\item{mtry}{Number of predictors (markers) that are randomly sampled
as candidates at each split. Will be passed on to the Random Forest
algorithm. Default is a third of the number of markers.}

\item{ntree}{Total number of trees. Defaults to 30000. More trees
lead to higher power.}

\item{npermut}{Number of different imputations that should
be created in the case of missing genotypes. ntree should
 be a multiple of npermut. Defaults to 100.}

\item{nodesize}{Minimum size of terminal nodes. Setting 
this number larger causes smaller trees to be grown 
(and thus take less time). Default is 5.}

\item{minTest}{Minimal number of slopes that need to be 
available for a marker pair in order for the splitA test 
to be applied. Default is 5 and should not be lowered for
statistical reasons.}

\item{nthreads}{Number of threads that should be used 
for building the forest.}

\item{fullTable}{logical; Should the output table include
mean slopes, selection frequencies, and ungrouped 
p-values? Normally not needed.}

\item{Rcutoff}{Absolute pearson correlation 
coefficient R (0 > Rcutoff <= 1) that should be used as a 
threshold for marker pairs to be excluded.}
}
\value{
Table with the marker pair indices and their 
 respective p-values from the pairedSF, the splitA, the selA, 
 and their ensemble. If there were multiple phenotypes, the
 results for each phenotype are returned as a list. If fullTable
 is TRUE, intermediate values, including the pair of mean
 slopes for each marker pair, the conditinal selection 
 frequencies, and ungrouped p-values.
}
\description{
function that detects epistasis between markers using random forest. 
For details see https://www.biorxiv.org/content/early/2018/06/21/353193
Note that all submitted markers in mappingData will be used to build the RF. 
The parameters markerInds1 and markerInds2, or pairMat can be used to
restrict interaction testing to a subset of marker pairs. If neither
of them are specified, all markers are tested.
}
