% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{rfMapperPar}
\alias{rfMapperPar}
\title{Parallel RF mapping}
\usage{
rfMapperPar(
  mappingData,
  nTrait = NULL,
  permutationGroups = NULL,
  file = NULL,
  permute = T,
  nPermutations = 1000,
  nforest,
  ntree,
  pMat = NULL,
  nCl,
  clType = "SOCK"
)
}
\arguments{
\item{mappingData}{A list containing all necessary information to 
map a trait as returned by preMap.}

\item{nTrait}{If the phenotype in mapping data is a matrix, nTrait
specifies which of the multiple traits provided, should be mapped.
Should be a single integer.}

\item{permutationGroups}{Optional; list with an entry for each 
sample indicating within which group it may .}

\item{file}{Optional; if the results are to be saved as a file,
this argument specifies the exact path.}

\item{permute}{Logical, should permutations be performed (if false
the real trait-values are used)?}

\item{nPermutations}{Number of permutations to be performed. If
permute is false, nPermutations is irrelevant.}

\item{nforest}{Number of randomizations for missing genotypes.}

\item{ntree}{Number of trees to grow per subforest.}

\item{pMat}{Optional; matrix containing predefined permutation 
schemes. The number of rows has to equal the number of samples 
and the number of columns has to equal the number of permutations.}

\item{nCl}{Integer; number of cores to use.}

\item{clType}{Character; type of cluster to create. 
Has to be a cluster-type implemented in snow.}
}
\value{
A matrix of scores with one column per permutation.
}
\description{
Parallelized implementation of trait mapping with Random Forest.
}
