% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RFQTLgenotype}
\alias{RFQTLgenotype}
\title{Fission yeast genotypes}
\format{
A matrix with 46 rows and 4481 columns: Each 
row corresponds to a strain, and each column to 
a genomic locus. Alleles are encoded as 0 or 1.
}
\source{
unpublished data.
}
\usage{
RFQTLgenotype
}
\description{
A matrix containing the genotypes of 
46 different fission yeast strains.
}
\keyword{datasets}
