% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RFQTLphenotype}
\alias{RFQTLphenotype}
\title{Fission yeast phenotypes}
\format{A matrix with 46 rows and 2 columns: Each 
row corresponds to a strain, and each column to 
a phenotype.}
\source{
unpublished data.
}
\usage{
RFQTLphenotype
}
\description{
A dataframe containing the phenotype of 
46 different fission yeast strains for two traits
}
\keyword{datasets}
