% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{treeStructure}
\alias{treeStructure}
\title{Reformat RF tree structure}
\usage{
treeStructure(treeMat)
}
\arguments{
\item{treeMat}{Matrix representing the structure of a tree of a forest.
Each row represents one node. The first two columns are the indices of 
the rows of the left and right child nodes, respectively. 
The third column indicates whether it is a terminal node 
or not (-3 means not terminal, -1 terminal, and 0 not used). 
The fourth column is the index in the genotype of the marker 
that was used for the split at this node. The last column is 
the mean out-of-bag (oob) trait value for this node.}
}
\value{
List of length 2. $nl: matrix with two columns containing
 indices of predictors. Predictors in the second column were used
 somewhere in the tree on the left side of the respective predictor
 in clumn a. $nr: same as $nl, but for pairs where predictors in
 column two were used somewhere on the right side of the predictors in column 1.
}
\description{
Function that reformats the structure of a RF tree for use in \code{RFepistasis}.
}
