% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{updateNAs}
\alias{updateNAs}
\title{Update missing values}
\usage{
updateNAs(markerVec, missing)
}
\arguments{
\item{markerVec}{Vector of indices of strains with missing information for
the current marker.}

\item{missing}{Vector of indices of strains with missing information for
the current trait.}
}
\value{
A list object of vectors containing the updated indices of strains
that miss the information for the current marker.
}
\description{
Adjust the indices of missing genotypes for missing phenotypes.
}
