% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{writeQTL}
\alias{writeQTL}
\title{Write .qtl file}
\usage{
writeQTL(
  QTLlist,
  traitNames = NULL,
  path,
  markerPositions = NULL,
  digits = NULL
)
}
\arguments{
\item{QTLlist}{QTL stored as a list-object as returned by QTLgrouper.}

\item{traitNames}{Optional; a vector with names for the traits 
which correspond to the rows in the p-value matrix used for QTLgrouper.}

\item{path}{Complete filepath ending in .qtl, specifying where 
the results should be saved.}

\item{markerPositions}{Optional; Matrix with three columns 
specifying the chromosome, start and end for each marker.}

\item{digits}{Optional; p-values are formatted to scientific 
convention and can be rounded according to digits.}
}
\value{
The output is written in a file, specified by path.
}
\description{
Writes a QTL-list to a .qtl file as a table.
}
