% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{combinePvalues}
\alias{combinePvalues}
\title{Fisher combine p-values}
\usage{
combinePvalues(p)
}
\arguments{
\item{p}{Numeric vector of p-values that should be combined}
}
\value{
Combined p-value.
}
\description{
Function that combines p-values using the fisher method. 
Ignores NA in the input. Returns NA is no valid p-values were
in the input (e.g., only NAs). If only one valid p-value is
submitted, this p-value is returned unchanged. Used in \code{RFepistasis}.
}
\examples{
p = c(0.01,0.4,0.02,NA)
combinePvalues(p)
}
