% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFQTL.R
\name{QTLgrouper}
\alias{QTLgrouper}
\title{Group QTL}
\usage{
QTLgrouper(pmat, sigThreshold, corThreshold, distThreshold, genotype, chrVec)
}
\arguments{
\item{pmat}{Matrix with p-values, traits (rows) X loci (columns).}

\item{sigThreshold}{Significance threshold to applied on the p-values.}

\item{corThreshold}{Threshold for absolute correlation values above
which signifcant loci are treated as linked to each other. If they 
are close they are joined to a single region, otherwise they are 
grouped together without including inbetween loci.}

\item{distThreshold}{Distance threshold up until which loci containing
highly correlated markers including all markers in between are linked.}

\item{genotype}{Allele-information for all samples (rows) and 
markers (columns).}

\item{chrVec}{Character-vector containing the chromosome on which
a given marker is located.}
}
\value{
List with QTLgroups. Each entry is a list of the target
 and a matrix with the QTL-regions involved.
}
\description{
Function that identifies significant marker-trait associations
and joins close markers to a single QTL if they are consecutive
and/or highly correlated and joins distant QTL to QTL-groups if
they contain highly correlated markers.
}
